Move `CollectionCreate` back to `MainNavbarContainer`
